@@ -1,86 +1,58 @@
 {
   "private": true,
-  "workspaces": {
-    "packages": [
-      "packages/*"
-    ],
-    "nohoist": [
-      "**/jest",
-      "**/jest/**",
-      "**/jasmine",
-      "**/jasmine/**",
-      "**/react",
-      "**/react/**",
-      "**/testing-library__jest-dom",
-      "**/testing-library__jest-dom/**",
-      "**/testing-library__react-hooks",
-      "**/testing-library__react-hooks/**",
-      "**/react-dom",
-      "**/react-dom/**",
-      "**/react-test-renderer",
-      "**/react-test-renderer/**"
-    ]
-  },
   "scripts": {
-    "build": "yarn lucide build && yarn lucide-react build && yarn lucide-preact build && yarn lucide-vue build && yarn lucide-vue-next build && yarn lucide-angular build",
-    "test": "yarn lucide build:icons && yarn lucide-react build:icons && yarn lucide-vue build:icons && jest",
-    "lucide": "yarn workspace lucide",
-    "lucide-angular": "yarn workspace lucide-angular",
-    "lucide-react": "yarn workspace lucide-react",
-    "lucide-preact": "yarn workspace lucide-preact",
-    "lucide-vue": "yarn workspace lucide-vue",
-    "lucide-vue-next": "yarn workspace lucide-vue-next",
-    "lucide-svelte": "yarn workspace lucide-svelte",
-    "lucide-static": "yarn workspace lucide-static",
-    "build:icons": "babel-node ./scripts/buildIcons.js --presets @babel/env",
-    "build:outline-icons": "babel-node ./scripts/outlineSvg.js --presets @babel/env",
-    "generate:supersprite": "babel-node ./scripts/generateSuperSVG.js",
-    "optimize": "esbuild  ./scripts/optimizeSvgs.js --bundle --platform=node | node",
-    "addtags": "babel-node ./scripts/addMissingKeysToTags.js --presets @babel/env",
-    "generate:changelog": "babel-node ./scripts/generateChangelog.js --presets @babel/env"
+    "build": "pnpm -r --filter './packages/**' build",
+    "test": "pnpm -r --filter './packages/**' test",
+    "lucide": "pnpm --filter lucide",
+    "lucide-angular": "pnpm --filter lucide-angular",
+    "lucide-react": "pnpm --filter lucide-react",
+    "lucide-react-native": "pnpm --filter lucide-react-native",
+    "lucide-preact": "pnpm --filter lucide-preact",
+    "lucide-solid": "pnpm --filter lucide-solid",
+    "lucide-vue": "pnpm --filter lucide-vue",
+    "lucide-vue-next": "pnpm --filter lucide-vue-next",
+    "lucide-svelte": "pnpm --filter lucide-svelte",
+    "lucide-static": "pnpm --filter lucide-static",
+    "build:icons": "node ./scripts/buildIcons.mjs --templateSrc ./packages/lucide/scripts/exportTemplate.mjs",
+    "build:outline-icons": "pnpm --filter outline-svg start",
+    "generate:supersprite": "node ./scripts/generateSuperSVG.mjs",
+    "optimize": "node ./scripts/optimizeSvgs.mjs",
+    "addjsons": "node scripts/addMissingIconJsonFiles.mjs",
+    "checkIcons": "node scripts/checkIconsAndCategories.mjs --presets @babel/env",
+    "tags2icons": "node scripts/migrateTagsToIcons.mjs --presets @babel/env",
+    "icons2tags": "node scripts/migrateIconsToTags.mjs --presets @babel/env",
+    "icons2categories": "node scripts/migrateIconsToCategories.mjs --presets @babel/env",
+    "categories2icons": "node scripts/migrateCategoriesToIcons.mjs --presets @babel/env",
+    "generate:changelog": "node ./scripts/generateChangelog.mjs",
+    "postinstall": "husky install",
+    "lint": "eslint --ext .ts,.js,.mjs ./{packages/lucide,scripts}",
+    "prepare": "husky install",
+    "gi": "node ./scripts/generate/generateIcons.mjs"
   },
   "devDependencies": {
-    "@ampproject/rollup-plugin-closure-compiler": "^0.25.2",
-    "@atomico/rollup-plugin-sizes": "^1.1.4",
-    "@babel/cli": "^7.10.5",
-    "@babel/core": "^7.11.1",
-    "@babel/node": "^7.13.10",
-    "@babel/plugin-transform-runtime": "^7.11.5",
-    "@babel/preset-env": "^7.11.0",
-    "@rollup/plugin-babel": "^5.0.0",
-    "@rollup/plugin-commonjs": "^17.1.0",
-    "@rollup/plugin-node-resolve": "^11.2.0",
-    "@rollup/plugin-replace": "^2.4.1",
-    "babel-jest": "^26.6.3",
-    "babel-plugin-add-import-extension": "^1.4.3",
-    "core-js": "3",
-    "esbuild": "^0.14.28",
-    "eslint": "^4.19.1",
-    "eslint-config-airbnb-base": "^12.1.0",
-    "eslint-config-prettier": "^2.9.0",
-    "eslint-plugin-import": "^2.5.0",
-    "eslint-plugin-prettier": "^2.5.0",
-    "husky": "^4.3.6",
-    "jest": "^26.4.2",
-    "lint-staged": "^10.5.3",
-    "minimist": "^1.2.6",
-    "node-fetch": "^2.6.7",
-    "prettier": "1.17.1",
-    "rollup": "^2.7.3",
-    "rollup-plugin-license": "^2.0.0",
-    "rollup-plugin-terser": "^5.2.0",
-    "rollup-plugin-visualizer": "^4.1.0",
-    "rollup-plugin-svelte": "^7.1.0",
-    "svg-outline-stroke": "^1.3.1",
-    "svgo": "^2.8.0",
-    "svgson": "^4.1.0"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
+    "eslint": "^8.36.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-prettier": "^4.2.1",
+    "husky": "^8.0.3",
+    "lint-staged": "^13.2.0",
+    "minimist": "^1.2.8",
+    "node-fetch": "^3.3.1",
+    "prettier": "2.7.1",
+    "svgo": "^3.0.2",
+    "svgson": "^5.2.1"
   },
   "lint-staged": {
-    "icons/*.svg": "npx babel-node ./scripts/optimizeStagedSvgs.js --presets @babel/env"
+    "icons/*.svg": "node ./scripts/optimizeStagedSvgs.mjs"
+  },
+  "packageManager": "pnpm@7.14.0",
+  "pnpm": {
+    "peerDependencyRules": {
+      "ignoreMissing": [
+        "@babel/core",
+        "@babel/preset-env"
+      ]
+    }
   }
 }
