@@ -1,7 +1,7 @@
 {
   "name": "lucide-angular",
   "description": "A Lucide icon library package for Angular applications",
-  "version": "0.48.0",
+  "version": "0.144.0",
   "author": "SMAH1",
   "license": "ISC",
   "homepage": "https://lucide.dev",
@@ -10,6 +10,9 @@
     "type": "git",
     "url": "https://github.com/lucide-icons/lucide.git",
     "directory": "packages/lucide-angular"
+  },
+  "publishConfig": {
+    "directory": "dist"
   },
   "keywords": [
     "Lucide",
@@ -22,30 +25,32 @@
     "Font Awesome"
   ],
   "scripts": {
-    "build": "yarn clean && yarn build:icons && yarn build:ng",
-    "clean": "npx shx rm -rf dist && npx shx rm -rf ./src/icons/*.ts",
-    "build:icons": "yarn --cwd ../../ build:icons --output=../packages/lucide-angular/src --templateSrc=../packages/lucide-angular/scripts/exportTemplate --iconFileExtention=.ts --exportFileName=index.ts",
-    "build:ng": "ng build --prod",
-    "test:headless": "ng test --no-watch --no-progress --browsers=ChromeHeadlessCI",
-    "test": "ng test",
+    "build": "pnpm clean && pnpm copy:license && pnpm build:icons && pnpm build:ng",
+    "copy:license": "cp ../../LICENSE ./LICENSE",
+    "clean": "rm -rf dist && rm -rf ./src/icons/*.ts",
+    "build:icons": "build-icons --output=./src --templateSrc=./scripts/exportTemplate.mjs --iconFileExtension=.ts --exportFileName=index.ts",
+    "build:ng": "ng build  --configuration production",
+    "test": "ng test --no-watch --no-progress --browsers=ChromeHeadlessCI",
+    "test:watch": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "postinstall": "ngcc"
+    "version": "pnpm version --git-tag-version=false"
   },
   "dependencies": {
-    "tslib": "^2.3.1"
+    "tslib": "^2.4.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1102.5",
-    "@angular/cli": "~11.2.15",
-    "@angular/common": "~11.2.14",
-    "@angular/compiler": "~11.2.14",
-    "@angular/compiler-cli": "~11.2.14",
-    "@angular/core": "~11.2.14",
-    "@angular/platform-browser": "~11.2.14",
-    "@angular/platform-browser-dynamic": "~11.2.14",
-    "@types/jasmine": "~3.10.2",
-    "@types/node": "^16.11.7",
+    "@angular-devkit/build-angular": "~14.2.6",
+    "@angular/cli": "~14.2.6",
+    "@angular/common": "~14.2.7",
+    "@angular/compiler": "~14.2.7",
+    "@angular/compiler-cli": "~14.2.7",
+    "@angular/core": "~14.2.7",
+    "@angular/platform-browser": "~14.2.7",
+    "@angular/platform-browser-dynamic": "~14.2.7",
+    "@lucide/build-icons": "workspace:*",
+    "@types/jasmine": "~4.3.0",
+    "@types/node": "^18.11.4",
     "codelyzer": "^6.0.2",
     "jasmine-core": "~3.10.1",
     "jasmine-spec-reporter": "~7.0.0",
@@ -54,11 +59,13 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-packagr": "^11.2.4",
+    "ng-packagr": "^14.2.1",
     "protractor": "~7.0.0",
     "puppeteer": "^8.0.0",
-    "ts-node": "~10.4.0",
+    "rxjs": "7.5.7",
+    "ts-node": "~10.9.1",
     "tslint": "~6.1.0",
-    "typescript": "~4.1.5"
+    "typescript": "~4.8.4",
+    "zone.js": "^0.11.8"
   }
 }
