@@ -1,7 +1,7 @@
 {
   "name": "lucide-react",
   "description": "A Lucide icon library package for React applications",
-  "version": "0.48.0",
+  "version": "0.144.0",
   "license": "ISC",
   "homepage": "https://lucide.dev",
   "bugs": "https://github.com/lucide-icons/lucide/issues",
@@ -18,25 +18,37 @@
   "unpkg": "dist/umd/lucide-react.min.js",
   "typings": "dist/lucide-react.d.ts",
   "sideEffects": false,
+  "files": [
+    "dist"
+  ],
   "scripts": {
-    "build": "yarn clean && yarn build:icons && (yarn build:es & yarn build:types & yarn build:bundles)",
-    "clean": "rm -rf dist && rm -rf stats && rm -rf ./src/icons/*.js",
-    "build:icons": "yarn --cwd ../../ build:icons --output=../packages/lucide-react/src --templateSrc=../packages/lucide-react/scripts/exportTemplate --renderUniqueKey",
-    "build:es": "babel src -d dist/esm",
-    "build:types": "yarn --cwd ../../ babel-node packages/lucide-react/scripts/buildTypes.js",
-    "build:bundles": "yarn --cwd ../../ rollup -c packages/lucide-react/rollup.config.js",
-    "test": "jest"
+    "build": "pnpm clean && pnpm copy:license && pnpm build:icons && pnpm typecheck && pnpm build:bundles && pnpm build:types",
+    "copy:license": "cp ../../LICENSE ./LICENSE",
+    "clean": "rm -rf dist && rm -rf stats && rm -rf ./src/icons/*.ts",
+    "build:icons": "build-icons --output=./src --templateSrc=./scripts/exportTemplate.mjs --renderUniqueKey --withAliases --aliasesFileExtension=.ts --iconFileExtension=.ts --exportFileName=index.ts",
+    "build:types": "node ./scripts/buildTypes.mjs",
+    "build:bundles": "rollup -c ./rollup.config.mjs",
+    "typecheck": "tsc",
+    "typecheck:watch": "tsc -w",
+    "test": "vitest run",
+    "version": "pnpm version --git-tag-version=false"
   },
   "devDependencies": {
+    "@lucide/build-icons": "workspace:*",
+    "@lucide/rollup-plugins": "workspace:*",
+    "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^11.2.6",
-    "babel-preset-react-app": "^10.0.0",
-    "jest": "^26.6.3",
-    "prop-types": "^15.7.2",
-    "react": "^16.5.1",
-    "react-dom": "^17.0.2"
+    "@types/prop-types": "^15.7.5",
+    "@types/react": "^18.0.21",
+    "@vitejs/plugin-react": "^2.1.0",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
+    "rollup": "^3.5.1",
+    "typescript": "^4.8.4",
+    "vite": "^3.1.0",
+    "vitest": "^0.24.3"
   },
   "peerDependencies": {
-    "prop-types": "^15.7.2",
     "react": "^16.5.1 || ^17.0.0 || ^18.0.0"
   }
 }
