@@ -1,7 +1,7 @@
 {
   "name": "lucide-svelte",
   "description": "A Lucide icon library package for Svelte applications",
-  "version": "0.48.0",
+  "version": "0.144.0",
   "license": "ISC",
   "homepage": "https://lucide.dev",
   "bugs": "https://github.com/lucide-icons/lucide/issues",
@@ -12,37 +12,46 @@
   },
   "author": "Eric Fennis",
   "amdName": "lucide-svelte",
-  "main": "dist/cjs/lucide-svelte.js",
-  "main:umd": "dist/umd/lucide-svelte.js",
+  "type": "module",
+  "main": "dist/esm/lucide-svelte.js",
   "module": "dist/esm/lucide-svelte.js",
-  "svelte": "dist/esm/lucide-svelte.js",
-  "unpkg": "dist/umd/lucide-svelte.min.js",
+  "svelte": "dist/svelte/lucide-svelte.js",
   "typings": "dist/lucide-svelte.d.ts",
   "sideEffects": false,
+  "files": [
+    "dist"
+  ],
   "scripts": {
-    "build": "yarn clean && yarn build:icons && (yarn build:es & yarn build:types & yarn build:bundles)",
+    "build": "pnpm clean && pnpm copy:license && pnpm build:icons && pnpm build:bundles && pnpm build:strip && pnpm build:types",
+    "copy:license": "cp ../../LICENSE ./LICENSE",
     "clean": "rm -rf dist && rm -rf stats && rm -rf ./src/icons/*.svelte && rm -f index.js",
-    "build:icons": "yarn --cwd ../../ build:icons --output=../packages/lucide-svelte/src --templateSrc=../packages/lucide-svelte/scripts/exportTemplate --exportFileName=index.js --iconFileExtention=.svelte --pretty=false",
-    "build:es": "babel src -d dist/esm --copy-files",
-    "build:types": "yarn --cwd ../../ babel-node packages/lucide-svelte/scripts/buildTypes.js",
-    "build:bundles": "yarn --cwd ../../ rollup -c packages/lucide-svelte/rollup.config.js",
-    "test": "jest",
-    "test:watch": "npm run test -- --watch"
+    "build:icons": "build-icons --output=./src --templateSrc=./scripts/exportTemplate.mjs --exportFileName=index.ts --iconFileExtension=.svelte --importImportFileExtension=.svelte --withAliases --aliasesFileExtension=.ts --aliasImportFileExtension=.svelte --pretty=false",
+    "build:types": "node ./scripts/buildTypes.mjs",
+    "build:bundles": "rollup -c ./rollup.config.mjs",
+    "build:strip": "svelte-strip strip src/ dist/svelte",
+    "test": "vitest run",
+    "version": "pnpm version --git-tag-version=false"
   },
   "devDependencies": {
+    "@lucide/rollup-plugins": "workspace:*",
+    "@lucide/build-icons": "workspace:*",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@sveltejs/vite-plugin-svelte": "^1.2.0",
     "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/preact": "^2.0.1",
     "@testing-library/svelte": "^3.0.3",
-    "babel-preset-preact": "^2.0.0",
-    "jest": "^26.6.3",
+    "@tsconfig/svelte": "^3.0.0",
+    "jsdom": "^20.0.3",
+    "rollup": "^3.5.1",
     "rollup-plugin-svelte": "^7.1.0",
-    "svelte": "^3.45.0",
-    "svelte-jester": "^2.3.1",
-    "svelte-preprocess": "^4.10.1",
-    "svelte2tsx": "^0.4.12"
+    "svelte": "^3.53.1",
+    "svelte-check": "^2.9.2",
+    "svelte-preprocess": "^4.10.7",
+    "svelte-strip": "^1.0.3",
+    "typescript": "^4.9.3",
+    "vite": "^3.1.0",
+    "vitest": "^0.23.2"
   },
   "peerDependencies": {
-    "svelte": "^3.45.0",
-    "svelte2tsx": "^0.4.12"
+    "svelte": "^3.49.0"
   }
 }
