@@ -1,6 +1,6 @@
 {
   "name": "lucide-vue-next",
-  "version": "0.48.0",
+  "version": "0.144.0",
   "author": "Eric Fennis",
   "description": "A Lucide icon library package for Vue 3 applications",
   "license": "ISC",
@@ -19,23 +19,33 @@
   "unpkg": "dist/umd/lucide-vue-next.min.js",
   "typings": "dist/lucide-vue-next.d.ts",
   "sideEffects": false,
+  "files": [
+    "dist",
+    "nuxt.js"
+  ],
   "scripts": {
-    "build": "yarn clean && yarn build:icons && (yarn build:es & yarn build:types & yarn build:bundles)",
-    "clean": "rm -rf dist && rm -rf ./src/icons/*.js",
-    "build:icons": "yarn --cwd ../../ build:icons --output=../packages/lucide-vue-next/src --templateSrc=../packages/lucide-vue-next/scripts/exportTemplate  --renderUniqueKey",
-    "build:es": "babel src -d dist/esm",
-    "build:types": "yarn --cwd ../../ babel-node packages/lucide-vue-next/scripts/buildTypes.js",
-    "build:bundles": "yarn --cwd ../../ rollup -c packages/lucide-vue-next/rollup.config.js",
-    "test": "jest",
-    "test:watch": "jest --watchAll"
+    "build": "pnpm clean && pnpm copy:license && pnpm build:icons && pnpm build:bundles && pnpm build:types",
+    "copy:license": "cp ../../LICENSE ./LICENSE",
+    "clean": "rm -rf dist && rm -rf ./src/icons/*.ts",
+    "build:icons": "build-icons --output=./src --templateSrc=./scripts/exportTemplate.mjs --renderUniqueKey --withAliases --aliasesFileExtension=.ts --iconFileExtension=.ts --exportFileName=index.ts",
+    "build:types": "node ./scripts/buildTypes.mjs",
+    "build:bundles": "rollup -c ./rollup.config.mjs",
+    "test": "vitest run",
+    "version": "pnpm version --git-tag-version=false"
   },
   "devDependencies": {
-    "@testing-library/vue": "^6.4.2",
-    "@vue/compiler-sfc": "3.0.1",
-    "@vue/test-utils": "2.0.0-rc.18",
-    "jest-serializer-vue": "^2.0.2",
-    "vue": "3.0.6",
-    "vue-jest": "^5.0.0-alpha.10"
+    "@lucide/rollup-plugins": "workspace:*",
+    "@lucide/build-icons": "workspace:*",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/vue": "^6.6.1",
+    "@vue/compiler-sfc": "3.2.45",
+    "@vue/test-utils": "2.2.4",
+    "@vitejs/plugin-vue": "^3.2.0",
+    "rollup": "^3.5.1",
+    "vite": "^3.1.0",
+    "vitest": "^0.24.3",
+    "vue": "3.2.45",
+    "jsdom": "^20.0.3"
   },
   "peerDependencies": {
     "vue": ">=3.0.1"
