@@ -1,6 +1,6 @@
 {
   "name": "lucide-vue",
-  "version": "0.48.0",
+  "version": "0.144.0",
   "author": "Eric Fennis",
   "description": "A Lucide icon library package for Vue 2 applications",
   "license": "ISC",
@@ -18,23 +18,33 @@
   "module": "dist/esm/lucide-vue.js",
   "unpkg": "dist/umd/lucide-vue.min.js",
   "sideEffects": false,
+  "files": [
+    "dist",
+    "nuxt.js"
+  ],
   "scripts": {
-    "build": "yarn clean && yarn build:icons && (yarn build:es & yarn build:bundles)",
+    "build": "pnpm clean && pnpm copy:license && pnpm build:icons && pnpm build:bundles",
+    "copy:license": "cp ../../LICENSE ./LICENSE",
     "clean": "rm -rf dist && rm -rf stats && rm -rf ./src/icons/*.js",
-    "build:icons": "yarn --cwd ../../ build:icons --output=../packages/lucide-vue/src --templateSrc=../packages/lucide-vue/scripts/exportTemplate",
-    "build:es": "babel src -d dist/esm",
-    "build:bundles": "yarn --cwd ../../ rollup -c packages/lucide-vue/rollup.config.js",
-    "test": "jest",
-    "test:watch": "jest --watchAll"
+    "build:icons": "build-icons --output=./src --templateSrc=./scripts/exportTemplate.mjs --renderUniqueKey --withAliases --aliasesFileExtension=.ts --iconFileExtention=.ts --exportFileName=index.ts",
+    "build:bundles": "rollup -c ./rollup.config.mjs",
+    "test": "vitest run",
+    "version": "pnpm version --git-tag-version=false"
   },
   "devDependencies": {
+    "@lucide/rollup-plugins": "workspace:*",
+    "@lucide/build-icons": "workspace:*",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/vue": "^5.8.2",
-    "@vue/test-utils": "^1.1.2",
+    "@vitejs/plugin-vue2": "^2.0.1",
+    "@vue/test-utils": "1.3.0",
     "jest-serializer-vue": "^2.0.2",
-    "vue": "^2.6.12",
-    "vue-jest": "^3.0.7",
-    "vue-template-compiler": "^2.6.12"
+    "rollup": "^3.5.1",
+    "typescript": "^4.8.4",
+    "vite": "^3.1.0",
+    "vitest": "^0.24.3",
+    "vue": "2.7.14",
+    "vue-template-compiler": "2.7.14"
   },
   "peerDependencies": {
     "vue": "^2.6.12"
