@@ -1,58 +1,72 @@
 {
+  "name": "@lucide/site",
   "private": true,
-  "name": "site",
   "version": "1.0.0",
   "author": "John Letey",
   "scripts": {
     "dev": "next dev",
     "copy-assets": "mkdir -p ./public/docs/images && cp -rf ../docs/images ./public/docs",
-    "build": "yarn copy-assets && next build",
+    "prebuild": "ts-node --swc scripts/preBuild.ts && pnpm -w lucide-react build",
+    "build": "pnpm copy-assets && pnpm prebuild && next build",
     "export": "next export -o build",
-    "deploy": "yarn build && yarn export",
+    "deploy": "pnpm build && pnpm export",
     "lint": "eslint .",
-    "test": "jest"
+    "lint:fix": "eslint --fix .",
+    "test": "jest",
+    "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@chakra-ui/react": "^1.6.0",
-    "@emotion/react": "^11",
-    "@emotion/styled": "^11",
+    "@chakra-ui/react": "1.8.8",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
     "@mdx-js/loader": "^1.6.22",
     "@mdx-js/react": "^1.6.22",
     "@next/mdx": "^11.0.0",
-    "@svgr/webpack": "^6.2.1",
+    "@svgr/webpack": "^6.3.1",
     "downloadjs": "^1.4.7",
-    "framer-motion": "^4",
+    "framer-motion": "^6.2.8",
+    "element-to-path": "^1.2.1",
     "fuse.js": "^6.5.3",
     "gray-matter": "^4.0.3",
     "js-yaml": "^4.1.0",
     "jszip": "^3.7.0",
     "lodash": "^4.17.20",
-    "lucide-react": "0.14.0",
+    "lucide-react": "workspace:*",
     "next": "12",
     "next-mdx-remote": "^3.0.2",
+    "object-path": "0.11.5",
     "prism-react-renderer": "^1.2.1",
-    "react": "^17.0.2",
+    "react": "17.0.2",
     "react-color": "^2.19.3",
-    "react-dom": "^17.0.2",
-    "react-svg-loader": "^3.0.3"
+    "react-dom": "17.0.2",
+    "react-svg-loader": "^3.0.3",
+    "svgson": "^5.2.1",
+    "svg-pathdata": "^6.0.3"
   },
   "devDependencies": {
-    "@next/eslint-plugin-next": "^11.1.0",
-    "@testing-library/dom": "^7.24.4",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.0.4",
-    "@testing-library/react-hooks": "^3.4.2",
+    "@next/eslint-plugin-next": "^12.2.5",
+    "@swc/core": "^1.3.35",
+    "@testing-library/dom": "^7.31.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^11.2.7",
+    "@testing-library/react-hooks": "^8.0.1",
+    "@types/jest": "^28.1.7",
     "@types/node": "^14.0.11",
     "@types/react": "^16.9.35",
     "@types/react-dom": "^16.9.8",
-    "@typescript-eslint/eslint-plugin": "^4.29.1",
-    "@typescript-eslint/parser": "^4.29.1",
+    "@typescript-eslint/eslint-plugin": "^5.34.0",
+    "@typescript-eslint/parser": "^5.34.0",
     "babel-jest": "^26.5.2",
     "babel-loader": "^8.1.0",
-    "cheerio": "^1.0.0-rc.3",
-    "eslint": "^7.32.0",
+    "eslint": "^8.22.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "jest": "^26.5.2",
+    "node-fetch": "2",
     "prettier": "^2.3.2",
+    "react-test-renderer": "17.0.2",
+    "ts-node": "~10.9.1",
+    "tslib": "^2.4.0",
     "typescript": "^4.3.5"
   }
 }
