@@ -1,8 +1,8 @@
-import { createReactComponent } from 'lucide-react'
-
 import IconButton from '../components/IconButton'
 import SearchInput from '../components/SearchInput'
+import createIconComponent from '../helpers/createIconComponent'
 import { Icon } from '../hooks/useSearch'
+import Skeleton from '../components/Skeleton/Skeleton'
 
 interface PageProps {
   query: string
@@ -25,16 +25,22 @@ const Icons = ({
         value={query}
         iconCount={icons.length}
         onChange={(event)  => setQuery(event.target.value)}
+        placeholder={icons.length ? `Search ${icons.length} icons`: 'Loading icons ..'}
       />
         <main>
           <div className='icon-grid'>
-            {searchResults.map(([name, iconNode] :any) => (
-              <IconButton
-                name={name}
-                key={name}
-                component={createReactComponent(name, iconNode)}
-              />
-            ))}
+            {icons.length ? (
+              searchResults.map(([name, iconNode]: any) => (
+                <IconButton
+                  name={name}
+                  key={name}
+                  component={createIconComponent(name, iconNode)}
+                />
+              ))
+            ) : (
+              <Skeleton />
+            )}
+
           </div>
           <footer>
             <a
