@@ -1,9 +1,12 @@
-<script>
+<script lang="ts">
   import defaultAttributes from './defaultAttributes'
-  export let name
+  import type { IconNode } from './types';
+
+  export let name: string
   export let color = 'currentColor'
-  export let size = 24
-  export let strokeWidth = 2
+  export let size: number|string = 24
+  export let strokeWidth: number|string = 2
+  export let iconNode: IconNode
 </script>
 
 <svg
@@ -15,5 +18,8 @@
   stroke-width={strokeWidth}
   class={`lucide-icon lucide lucide-${name} ${$$props.class ?? ''}`}
 >
+  {#each iconNode as [tag, attrs]}
+    <svelte:element this={tag} {...attrs}/>
+  {/each}
   <slot />
 </svg>
