@@ -1,4 +1,4 @@
-<script>
+<script lang="ts">
   import Smile from '../src/icons/smile.svelte'
 </script>
 
