@@ -1,70 +1,81 @@
-import { Button, ButtonProps, Flex, Text, useToast } from '@chakra-ui/react';
+import { Button, ButtonProps, Tooltip, useToast } from '@chakra-ui/react';
 import download from 'downloadjs';
-import copy from 'copy-to-clipboard';
-import { memo } from 'react';
+import { memo, useCallback } from 'react';
+import { createLucideIcon, IconNode } from 'lucide-react';
 import { useCustomizeIconContext } from './CustomizeIconContext';
-import { IconWrapper } from './IconWrapper';
+import { useRouter } from 'next/router';
 
-interface IconListItemProps {
+interface IconListItemProps extends ButtonProps {
   name: string;
-  content: string;
-  contributors: any[];
-  src: string;
-  onClick?: ButtonProps['onClick'];
+  iconNode: IconNode;
 }
 
-const IconListItem = ({ name, content, onClick, src: svg }: IconListItemProps) => {
+const IconListItem = ({ name, iconNode }: IconListItemProps) => {
+  const router = useRouter()
   const toast = useToast();
   const { color, size, strokeWidth, iconsRef } = useCustomizeIconContext();
 
+  const Icon = createLucideIcon(name, iconNode)
+
+  const handleClick:ButtonProps['onClick'] = useCallback(async (event) => {
+    const src = (iconsRef.current[name].outerHTML).replace(/(\r\n|\n|\r|(>\s\s<))/gm, "")
+
+    if (event.shiftKey) {
+      await navigator.clipboard.writeText(src)
+
+      toast({
+        title: 'Copied!',
+        description: `Icon "${name}" copied to clipboard.`,
+        status: 'success',
+        duration: 1500,
+      });
+      return
+    }
+    if (event.altKey) {
+      download(src, `${name}.svg`, 'image/svg+xml');
+      return
+    }
+
+    router.push({
+      pathname: `/icon/${name}`,
+      query: (
+        router.query?.search != null
+        ? { search: router.query.search }
+        : {}
+      ),
+    },
+    undefined,
+    {
+      shallow: true,
+      scroll: false
+    })
+  }, [iconsRef, name, router, toast])
+
+
   return (
-    <Button
-      variant="ghost"
-      borderWidth="1px"
-      rounded="lg"
-      padding={2}
-      height={32}
-      position="relative"
-      whiteSpace="normal"
-      onClick={event => {
-        const src = iconsRef.current[name].outerHTML ?? svg
-        if (event.shiftKey) {
-          copy(src);
-          toast({
-            title: 'Copied!',
-            description: `Icon "${name}" copied to clipboard.`,
-            status: 'success',
-            duration: 1500,
-          });
-        }
-        if (event.altKey) {
-          download(src, `${name}.\svg`, 'image/svg+xml');
-        }
-        if (onClick) {
-          onClick(event);
-        }
-      }}
-      key={name}
-      alignItems="center"
-    >
-      <Flex direction="column" align="center" justify="stretch" width="100%" gap={4}>
-        <Flex flex={2} flexBasis="100%" minHeight={10} align="flex-end">
-          <IconWrapper
-            content={content}
-            stroke={color}
-            strokeWidth={strokeWidth}
-            height={size}
-            width={size}
-            ref={iconEl => (iconsRef.current[name] = iconEl)}
-          />
-        </Flex>
-        <Flex flex={1} minHeight={10} align="center">
-          <Text wordBreak="break-word" maxWidth="100%">
-            {name}
-          </Text>
-        </Flex>
-      </Flex>
-    </Button>
+    <Tooltip label={name}>
+      <Button
+        as="a"
+        variant="ghost"
+        borderWidth="1px"
+        rounded="lg"
+        padding={2}
+        height={20}
+        width={20}
+        position="relative"
+        whiteSpace="normal"
+        alignItems="center"
+        sx={{cursor: 'pointer'}}
+        onClick={handleClick}
+      >
+        <Icon
+          ref={iconEl => (iconsRef.current[name] = iconEl)}
+          size={size}
+          stroke={color}
+          strokeWidth={strokeWidth}
+        />
+      </Button>
+    </Tooltip>
   );
 };
 
