@@ -1,7 +1,7 @@
 import { Box, Tooltip, useColorMode } from "@chakra-ui/react";
 import theme from '../lib/theme';
 
-const ModifiedTooltip = ({}) => {
+const ModifiedTooltip = () => {
   const { colorMode } = useColorMode();
 
   return (
