@@ -6,11 +6,19 @@ import IconDetailOverlay from '../components/IconDetailOverlay';
 import { useRouter } from 'next/router';
 import Header from '../components/Header';
 import MobileMenu from '../components/MobileMenu';
-import { useEffect, useMemo } from 'react';
+import { useMemo } from 'react';
+import { GetStaticPropsResult, NextPage } from 'next';
+import { IconEntity, Category } from '../types';
+import { getAllCategories } from 'src/lib/categories';
 
-const IndexPage = ({ data }) => {
+interface HomePageProps {
+  data: IconEntity[]
+  categories: Category[]
+}
+
+const HomePage: NextPage<HomePageProps> = ({ data, categories }) => {
   const router = useRouter();
-  const getIcon = iconName => data.find(({ name }) => name === iconName) || {};
+  const getIcon = iconName => data.find(({ name }) => name === iconName);
 
   const currentIcon = useMemo(() => {
     return getIcon(router.query.iconName)
@@ -22,22 +30,30 @@ const IndexPage = ({ data }) => {
       <IconDetailOverlay
         open={!!currentIcon?.name}
         icon={currentIcon}
-        close={() => router.push('/', undefined, { shallow: true })}
+        close={() => router.push({
+          pathname: '/icon/[iconName]',
+          query: {
+            ...router.query,
+            iconName: '',
+          },
+        }, undefined, { shallow: true })}
       />
       <Header {...{ data }} />
-      <IconOverview {...{ data }} />
+      <IconOverview {...{ data, categories }} />
     </Layout>
   );
 };
 
-export async function getStaticProps() {
-  const data = await getAllData();
+export async function getStaticProps(): Promise<GetStaticPropsResult<HomePageProps>> {
+  const data = await getAllData({ withChildKeys: true });
+  const categories = await getAllCategories()
 
   return {
     props: {
       data,
+      categories,
     },
   };
 }
 
-export default IndexPage;
+export default HomePage;
