@@ -5,9 +5,9 @@ import MobileMenu from '../components/MobileMenu';
 import { Box, Heading, Text } from '@chakra-ui/react';
 import { promises as fs } from 'fs';
 import { resolve } from 'path';
-import { GetStaticProps, NextPage } from 'next';
+import { GetStaticProps } from 'next';
 
-const LicensePage = ({ licenseText }) => {
+const LicensePage = ({ licenseText }: { licenseText: string[] }): JSX.Element => {
   return (
     <Box>
       <MobileMenu />
@@ -43,12 +43,17 @@ const LicensePage = ({ licenseText }) => {
 
 export default LicensePage;
 
-export const getStaticProps: GetStaticProps = async ({ params }) => {
+export const getStaticProps: GetStaticProps = async () => {
   const doc: string = await fs.readFile(resolve('../LICENSE'), 'utf-8');
 
   const licenseText = doc
     .split(/\n{2,}/)
-    .map(paragraph => paragraph.split('\n').join(' ').trim())
+    .map(paragraph =>
+      paragraph
+        .split('\n')
+        .join(' ')
+        .trim(),
+    )
     .filter(Boolean);
 
   return { props: { licenseText } };
