@@ -2,12 +2,19 @@ import Layout from '../../components/Layout';
 import HeadingNavigationProvider from '../../components/HeadingNavigationProvider';
 import MobileMenu from '../../components/MobileMenu';
 import { Stack } from '@chakra-ui/react';
-import Package from '../../components/Package';
-import packageData from '../../lib/packageData';
+import Package, { PackageItem } from '../../components/Package';
+import packagesData from '../../data/packageData.json';
+import thirdPartyPackagesData from '../../data/packageData.thirdParty.json';
 import { Heading } from '@chakra-ui/react';
 import fetchPackages from '../../lib/fetchPackages';
+import { GetStaticPropsResult } from 'next';
 
-const PackagesPage = ({ packages }) => {
+interface PackagesPageProps {
+  packages: PackageItem[]
+  thirdPartyPackages: PackageItem[]
+}
+
+const PackagesPage = ({ packages, thirdPartyPackages }: PackagesPageProps): JSX.Element => {
   return (
     <HeadingNavigationProvider>
       <MobileMenu />
@@ -16,7 +23,18 @@ const PackagesPage = ({ packages }) => {
           Packages
         </Heading>
         <Stack spacing={8} align="center">
-          {packages.length ? packages.map(packageItem => <Package {...packageItem} />) : null}
+          {packages.length
+            ? packages.map((packageItem) => <Package key={packageItem.name} {...packageItem} />)
+            : null}
+        </Stack>
+
+        <Heading as="h1" marginBottom={6} marginTop={12} textAlign="center">
+          Third party packages
+        </Heading>
+        <Stack spacing={8} marginBottom={6} align="center">
+          {thirdPartyPackages.length
+            ? thirdPartyPackages.map((packageItem) => (<Package key={packageItem.name} {...packageItem} />))
+            : null}
         </Stack>
       </Layout>
     </HeadingNavigationProvider>
@@ -25,23 +43,23 @@ const PackagesPage = ({ packages }) => {
 
 export default PackagesPage;
 
-export async function getStaticProps({ params }) {
-  const packages = (await fetchPackages())
+export async function getStaticProps(): Promise<GetStaticPropsResult<PackagesPageProps>> {
+  const packages: PackageItem[] = (await fetchPackages())
     .filter(Boolean)
-    .filter(packageObj => !packageObj.private)
-    .map(({ name, description, flutter = false }) => {
+    .filter(packageObj => !packageObj.private && packageObj.name in packagesData)
+    .map(({ name, description, flutter }) => {
       const packageDirectory = flutter ? 'lucide-flutter' : name;
 
       return {
         name,
         description,
-        image: `/package-logos/${packageDirectory}-small.svg`,
-        source: `https://github.com/lucide-icons/lucide/tree/master/packages/${packageDirectory}`,
+        source: `https://github.com/lucide-icons/lucide/tree/main/packages/${packageDirectory}`,
         documentation: `/docs/${packageDirectory}`,
-        ...packageData[packageDirectory],
+        ...packagesData[packageDirectory],
+        icon: `/framework-logos/${packagesData[packageDirectory].icon}.svg`,
       };
     })
     .sort((a, b) => a.order - b.order);
 
-  return { props: { packages } };
+  return { props: { packages, thirdPartyPackages: thirdPartyPackagesData } };
 }
